pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bookocr"
version = "1.0.2"
authors = [
  { name="Artem Poliakov", email="artem4250@gmail.com" },
]
description = "Optical character recognition (OCR) tool for printed book pages"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "opencv-python>=4.5.5.64",
    "numpy>=1.23.5",
    "pandas>=2.0.1",
    "scipy>=1.10.1",
    "tensorflow>=2.12.0",
    "Pillow>=9.5.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
bookocr = ["*.h5"]

[project.urls]
"Homepage" = "https://github.com/Artem259/bookocr"